Added the Constraints.py file and introduced HVDC constraint test
The HVDC constraint test takes three values, two energy prices and a reserve price
and compares them to see if a reserve constraint was likely in that period.
It has support for both absolute and relative tolerances as well as a minimum price
split in order to determine the non-zero prices which cloud the analysis.