Added support for various helper "masks" to make splitting data easier
The masks function is applied to a DataFrame and is used to apply general masks
to a DataFrame. Currently has support for all of the major boolean constraints
(eq, lt, le, gt, ge, ne) and is applied in a single function.
These methods may then be called using a key value combination along with the appropriate
mask rather than the initial bulky pandas syntax of chaining booleans