Fixed the hydrology datasets for those ugly spikes...
There appears to be a bug in the NIWA database when you select a series
table. This bug introduces a range of data spikes which shouldn't exist
and completely ruins any data analysis which could be done.
However, if you run it as a Series Chart this problem appears to be removed,
One problem with this is that it then stores the date in an ugly format.

I've modified the functions to accept "niwa_date" as an argument.
This will then use the custom date formatting in order to switch it to
a python date time object