Added a function to load the entirety of the IL data as a joint dataframe.
This required creation of a new function, create_il_dataset, which takes
a range of optional arguments and creates the dataset.
Currently, it has smart defaults set and should just be able to run without
modifying the KWD arguments.
Also has the ability to map the dataset and apply a time index simply to it

Also fixed a bug which was in the load map data function,
this consisted of changing the folder name location