Began a major overhaul of the NZEM module to make it more useful to me personally.
This is part of an ongoing process to make it more useful, pythonic and cleaner as
it is currently a bit of a mess.

This update adds the wits module, which is utilised to scrape the free to air wits
site for offers and demand data. It is still a work in progress at this stage