Added a method to conduct a "Naive" clearing analysis.
This analysis uses the offer stacks, as well as a requirement for capacity.
It sorts the stacks by price, removes non zero offers and clears the solution.
It does not take into account any form of constraints at all.

It does some preliminary post processing to determine the clearing
quantity of each unit, with support for partial dispatch of the marginal
unit.

It also determines the price.

If clearing an energy offer it will not take into account losses or
transmission constraints.
Instead it will clear the country as a single stack limiting usefulness
slightly